Move schemas to folder
